Fixed merge conflicts + organized inventory folder
- Fixed merge conflicts
- Moved barcode and manual-entry into their own folders within inventory
