+++ client/src/app/inventory/barcode/barcode-print-dialog.html
