+++ client/src/app/inventory/barcode/barcode-print-dialog.ts
@@ -1,7 +1,7 @@
 import { Component, inject } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogModule } from "@angular/material/dialog";
 import { MatCheckbox } from "@angular/material/checkbox";
-import { Inventory } from './inventory';
+import { Inventory } from '../inventory';
 import { CommonModule } from '@angular/common';
 import { MatDialogRef } from '@angular/material/dialog';
 import { MatButtonModule } from '@angular/material/button';
