+++ client/src/app/inventory/barcode/barcode-print-quantity-dialog.html
