+++ client/src/app/inventory/barcode/barcode-print-quantity-dialog.scss
