+++ client/src/app/inventory/barcode/barcode-print-window.service.ts
