+++ client/src/app/inventory/manual-entry/manual-entry.component.scss
