+++ client/src/app/inventory/manual-entry/manual-entry.html
